ajouter les exception de les fonctions de delete de faculty et matiér…

+++ backend/src/faculties/faculties.service.ts
@@ -38,6 +38,10 @@ export class FacultiesService {
   }
 
   async remove(id: number) {
-    return await this.facultyRepository.delete(id);
+    const faculty = await this.facultyRepository.findOne({where:{id}});
+    if (!faculty) {
+      throw new NotFoundException(`Faculty #${id} not found`);
+    }
+    return  this.facultyRepository.delete(id);
   }
 }
